Implement ValueSetLookup controller
#12553 - As a developer, will have created a value set API that is componentized and integrates with C2S architecture

+++ vss/src/main/java/gov/samhsa/c2s/vss/domain/ValueSetCategory.java
@@ -21,7 +21,7 @@ public class ValueSetCategory {
 
     private boolean isFederal;
 
-    private Long displayOrder;
+    private int displayOrder;
 
     @OneToMany(mappedBy = "valueSetCategory")
     @NotAudited
